refactor: LottoResult, WinningNumber 패키지 이동
LottoResult, WinningNumber를 dto 하위 패키지로 이동
