+++ src/main/java/lotto/model/service/LottoChecker.java
@@ -3,8 +3,8 @@ package lotto.model.service;
 import java.util.List;
 import lotto.model.domain.Lotto;
 import lotto.model.domain.LottoRank;
-import lotto.model.domain.LottoResult;
-import lotto.model.domain.WinningNumber;
+import lotto.model.dto.LottoResult;
+import lotto.model.dto.WinningNumber;
 
 public class LottoChecker {
