+++ src/test/java/lotto/view/output/OutputViewTest.java
@@ -1,0 +1,5 @@
+package lotto.view.output;
+
+class OutputViewTest {
+
+}
